gamelib: build every HTTP client in one place
The Windows TLS fix named the provider in `openretro::agent` and missed
the other constructor: `support::fetch`, which downloads the WHDLoad and
Soft-Kicker archives, built its own. ureq defaults to Rustls whatever is
compiled in -- its own documentation says of the alternative that "the
setting is never picked up automatically" -- so on Windows that download
panicked mid-request:

    uri scheme is https, provider is Rustls but feature is not enabled: rustls

which is the first thing a new Windows install does.

There is now one constructor, `gamelib::http::agent`, and both callers go
through it. A test reads the sibling sources and fails if a second
`config_builder` appears, because this is a mistake that cannot be caught
by running the code: an agent built by hand works perfectly on the
machine of whoever writes it and panics on somebody else's. Run against
the tree before this change, it fails on both files.
